scripts/requirements.txt
pandas
requests
python-dotenv
numpy